zn score tests move
